backend/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Marcus Schiesser <mail@marcusschiesser.de>" ]
readme = "README.md"

[tool.poetry.scripts]
generate = "app.engine.generate:generate_datasource"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.109.1"
python-dotenv = "^1.0.0"
aiostream = "^0.5.2"
cachetools = "^5.3.3"
llama-index = "^0.11.17"
llama-index-indices-managed-llama-cloud = "^0.3.1"
docx2txt = "^0.8"
xhtml2pdf = "^0.2.14"
markdown = "^3.7"
e2b_code_interpreter = "0.0.11b38"
llama-index-utils-workflow = "^0.2.2"
ipykernel = "^6.29.5"
llama-index-readers-web = "^0.2.4"
tavily-python = "^0.5.0"
llama-hub = "^0.0.79.post1"
firecrawl-py = "^1.4.0"
sendgrid = "^6.11.0"
crawl4ai = {extras = ["transformers"], version = "^0.3.72"}
pydub = "^0.25.1"
llama-index-llms-nvidia = "^0.2.6"
llama-index-embeddings-nvidia = "^0.2.4"
google-search-results = "^2.4.2"
setuptools = "^80.9.0"
cerebras-cloud-sdk = "^1.50.1"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[tool.poetry.group]
[tool.poetry.group.dev]
[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"

[tool.mypy]
python_version = "3.11"
plugins = "pydantic.mypy"
exclude = [ "tests", "venv", ".venv", "output", "config" ]
check_untyped_defs = true
warn_unused_ignores = false
show_error_codes = true
namespace_packages = true
ignore_missing_imports = true
follow_imports = "silent"
implicit_optional = true
strict_optional = false
disable_error_code = [ "return-value", "assignment" ]

[[tool.mypy.overrides]]
module = "app.*"
ignore_missing_imports = false

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"